refactor(journey): extract shared Clock component
Pull the ticking screen clock out of DeparturesBoard into its own
component with an opt-in seconds prop, so the upcoming TerminalHeader
can reuse it.
